Hw 29

+++ src/main/java/com/internet/shop/controller/user/DeleteUserController.java
@@ -1,6 +1,7 @@
 package com.internet.shop.controller.user;
 
 import com.internet.shop.lib.Injector;
+import com.internet.shop.model.Role;
 import com.internet.shop.service.ShoppingCartService;
 import com.internet.shop.service.UserService;
 import java.io.IOException;
@@ -22,11 +23,15 @@ public class DeleteUserController extends HttpServlet {
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
         Long userId = Long.valueOf(req.getParameter("id"));
-        shoppingCartService.delete(shoppingCartService.getByUserId(userId));
+        if (userService.get(userId).getRoles().stream()
+                .map(Role::getRoleName)
+                .noneMatch(r -> r.equals(Role.RoleName.ADMIN))) {
+            shoppingCartService.delete(shoppingCartService.getByUserId(userId));
+        }
         userService.delete(userId);
         HttpSession session = req.getSession();
         if (userId.equals(session.getAttribute(LoginController.USER_ID))) {
-            session.setAttribute(LoginController.USER_ID, null);
+            session.invalidate();
             resp.sendRedirect(req.getContextPath() + "/user/registration");
         }
         resp.sendRedirect(req.getContextPath() + "/user/all");
